requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
numpy>=2.2.5
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0
requests>=2.26.0
python-dotenv>=0.19.0
matplotlib>=3.4.2
seaborn>=0.11.1
websockets>=10.3
openai>=0.27.0
newsapi-python>=0.2.6